streamlit/requirements.txt
streamlit
google-cloud-aiplatform
google-auth
av
Pillow
vertexai
requests
python-dotenv
